feat: Support for silero-vad v5 model on ios
